+++ src/ng2-smart-table/components/cell/cell-view-mode/custom-view.component.ts
@@ -1,0 +1,43 @@
+import {
+  Component,
+  Input,
+  ComponentFactoryResolver,
+  ViewChild,
+  ViewContainerRef,
+  OnInit,
+  OnDestroy,
+} from '@angular/core';
+
+import { Cell } from '../../../lib/data-set/cell';
+
+@Component({
+  selector: 'custom-view-component',
+  template: `
+    <ng-template #dynamicTarget></ng-template>
+  `,
+})
+export class CustomViewComponent implements OnInit, OnDestroy {
+
+  customComponent: any;
+  @Input() cell: Cell;
+  @ViewChild('dynamicTarget', { read: ViewContainerRef }) dynamicTarget: any;
+
+  constructor(private resolver: ComponentFactoryResolver) {
+  }
+
+  ngOnInit() {
+    if (this.cell && !this.customComponent) {
+      const componentFactory = this.resolver.resolveComponentFactory(this.cell.getColumn().renderComponent);
+      this.customComponent = this.dynamicTarget.createComponent(componentFactory);
+
+      // set @Inputs and @Outputs of custom component
+      this.customComponent.instance.value = this.cell.getValue();
+    }
+  }
+
+  ngOnDestroy() {
+    if (this.customComponent) {
+      this.customComponent.destroy();
+    }
+  }
+}