+++ src/ng2-smart-table/components/cell/cell.component.ts
@@ -1,54 +1,38 @@
-import { Component, Input, Output, EventEmitter, ViewChild, ElementRef } from '@angular/core';
+import { Component, Input, Output, EventEmitter } from '@angular/core';
 
+import { Grid } from '../../lib/grid';
 import { Cell } from '../../lib/data-set/cell';
+import { Row } from '../../lib/data-set/row';
 
 @Component({
   selector: 'ng2-smart-table-cell',
-  styles: [require('./cell.scss')],
   template: `
-    <div #cellContainer *ngIf="!cell.getRow().isInEditing && cell.getColumn().type !== 'html'">{{ cell.getValue() }}</div>
-    <div #cellContainer *ngIf="!cell.getRow().isInEditing && cell.getColumn().type === 'html'" [innerHTML]="cell.getValue()"></div>
-    <input *ngIf="cell.getRow().isInEditing" 
-      [ngClass]="inputClass"
-      class="form-control"
-      [(ngModel)]="cell.newValue"
-      [name]="cell.getColumn().id" 
-      [placeholder]="cell.getColumn().title"
-      [disabled]="!cell.getColumn().isEditable"
-      (click)="onClick($event)"
-      (keydown.enter)="onEdited($event)" 
-      (keydown.esc)="onStopEditing()">
-  `
+    <table-cell-view-mode *ngIf="!isInEditing" [cell]="cell"></table-cell-view-mode>
+    <table-cell-edit-mode *ngIf="isInEditing" [cell]="cell"
+                          [inputClass]="inputClass"
+                          (edited)="onEdited($event)">
+    </table-cell-edit-mode>
+  `,
 })
 export class CellComponent {
 
+  @Input() grid: Grid;
+  @Input() row: Row;
+  @Input() editConfirm: EventEmitter<any>;
+  @Input() createConfirm: EventEmitter<any>;
+  @Input() isNew: boolean;
   @Input() cell: Cell;
   @Input() inputClass: string = '';
   @Input() mode: string = 'inline';
+  @Input() isInEditing: boolean = false;
 
-  @Output() public edited: EventEmitter<any> = new EventEmitter<any>();
+  @Output() edited = new EventEmitter<any>();
 
-  @ViewChild('cellContainer') cellRef: ElementRef;
-
-  onStopEditing(): boolean {
-    this.cell.getRow().isInEditing = false;
-    return false;
-  }
-
-  ngAfterViewInit(): void {
-    const cellRenderFunc = this.cell.getColumn().getCellRenderFunction();
-
-    if (cellRenderFunc) {
-      cellRenderFunc.call(null, this.cell, this.cellRef.nativeElement)
+  onEdited(event: any) {
+    if (this.isNew) {
+      this.grid.create(this.grid.getNewRow(), this.createConfirm);
+    } else {
+      this.grid.save(this.row, this.editConfirm);
     }
-  }
-
-  onEdited(event): boolean {
-    this.edited.emit(event);
-    return false;
-  }
-
-  onClick(event): void {
-    event.stopPropagation();
   }
 }