+++ src/ng2-smart-table/components/pager/pager.component.ts
@@ -1,50 +1,52 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
 
 import { DataSource } from '../../lib/data-source/data-source';
 
 @Component({
   selector: 'ng2-smart-table-pager',
-  styles: [require('./pager.scss')],
+  styleUrls: ['./pager.component.scss'],
   template: `
-  <nav *ngIf="shouldShow()" class="ng2-smart-pagination-nav">
-    <ul class="ng2-smart-pagination pagination">
-      <li class="ng2-smart-page-item page-item" [ngClass]="{disabled: getPage() == 1}">
-        <a class="ng2-smart-page-link page-link" href="#" 
-        (click)="getPage() == 1 ? false : paginate(1)" aria-label="First">
-          <span aria-hidden="true">&laquo;</span>
-          <span class="sr-only">First</span>
-        </a>
-      </li>
-      <li class="ng2-smart-page-item page-item" 
-      [ngClass]="{active: getPage() == page}" *ngFor="let page of getPages()">
-        <span class="ng2-smart-page-link page-link" 
-        *ngIf="getPage() == page">{{ page }} <span class="sr-only">(current)</span></span>
-        <a class="ng2-smart-page-link page-link" href="#" 
-        (click)="paginate(page)" *ngIf="getPage() != page">{{ page }}</a>
-      </li>
+    <nav *ngIf="shouldShow()" class="ng2-smart-pagination-nav">
+      <ul class="ng2-smart-pagination pagination">
+        <li class="ng2-smart-page-item page-item" [ngClass]="{disabled: getPage() == 1}">
+          <a class="ng2-smart-page-link page-link" href="#"
+          (click)="getPage() == 1 ? false : paginate(1)" aria-label="First">
+            <span aria-hidden="true">&laquo;</span>
+            <span class="sr-only">First</span>
+          </a>
+        </li>
+        <li class="ng2-smart-page-item page-item"
+        [ngClass]="{active: getPage() == page}" *ngFor="let page of getPages()">
+          <span class="ng2-smart-page-link page-link"
+          *ngIf="getPage() == page">{{ page }} <span class="sr-only">(current)</span></span>
+          <a class="ng2-smart-page-link page-link" href="#"
+          (click)="paginate(page)" *ngIf="getPage() != page">{{ page }}</a>
+        </li>
 
-      <li class="ng2-smart-page-item page-item" 
-      [ngClass]="{disabled: getPage() == getLast()}">
-        <a class="ng2-smart-page-link page-link" href="#" 
-        (click)="getPage() == getLast() ? false : paginate(getLast())" aria-label="Last">
-          <span aria-hidden="true">&raquo;</span>
-          <span class="sr-only">Last</span>
-        </a>
-      </li>
-    </ul>
-  </nav>
-  `
+        <li class="ng2-smart-page-item page-item"
+        [ngClass]="{disabled: getPage() == getLast()}">
+          <a class="ng2-smart-page-link page-link" href="#"
+          (click)="getPage() == getLast() ? false : paginate(getLast())" aria-label="Last">
+            <span aria-hidden="true">&raquo;</span>
+            <span class="sr-only">Last</span>
+          </a>
+        </li>
+      </ul>
+    </nav>
+  `,
 })
-export class PagerComponent {
+export class PagerComponent implements OnInit {
 
   @Input() perPage: number;
   @Input() source: DataSource;
+
+  @Output() changePage = new EventEmitter<any>();
 
   protected pages: Array<any>;
   protected page: number;
   protected count: number = 0;
 
-  ngOnInit(): void {
+  ngOnInit() {
     this.source.onChanged().subscribe((changes) => {
       this.page = this.source.getPaging().page;
       this.count = this.source.count();
@@ -64,7 +66,7 @@ export class PagerComponent {
    * if a new element was added to the beginning of the table - then to the first page
    * @param changes
    */
-  processPageChange(changes): void {
+  processPageChange(changes: any) {
     if (changes['action'] === 'prepend') {
       this.source.setPage(1);
     }
@@ -80,6 +82,7 @@ export class PagerComponent {
   paginate(page: number): boolean {
     this.source.setPage(page);
     this.page = page;
+    this.changePage.emit({ page });
     return false;
   }
 
@@ -95,12 +98,12 @@ export class PagerComponent {
     return Math.ceil(this.count / this.perPage);
   }
 
-  protected isPageOutOfBounce(): boolean {
+  isPageOutOfBounce(): boolean {
     return (this.page * this.perPage) >= (this.count + this.perPage) && this.page > 1;
   }
 
-  protected initPages() {
-    let pagesCount = this.getLast();
+  initPages() {
+    const pagesCount = this.getLast();
     let showPagesCount = 4;
     showPagesCount = pagesCount < showPagesCount ? pagesCount : showPagesCount;
     this.pages = [];
@@ -113,7 +116,7 @@ export class PagerComponent {
       let lastOne = middleOne + Math.floor(showPagesCount / 2);
       lastOne = lastOne >= pagesCount ? pagesCount : lastOne;
 
-      let firstOne = lastOne - showPagesCount + 1;
+      const firstOne = lastOne - showPagesCount + 1;
 
       for (let i = firstOne; i <= lastOne; i++) {
         this.pages.push(i);