+++ src/ng2-smart-table/lib/data-set/column.ts
@@ -2,46 +2,60 @@ import { DataSet } from './data-set';
 
 export class Column {
 
-  public title: string = '';
-  public type: string = '';
-  public class: string = '';
-  public isSortable: boolean = false;
-  public isEditable: boolean = true;
-  public isFilterable: boolean = false;
-  public sortDirection: string = '';
-  public defaultSortDirection: string = '';
-  protected compareFunction: Function;
-  protected valuePrepareFunction: Function;
-  protected filterFunction: Function;
-  protected cellRenderFunction: Function;
+  title: string = '';
+  type: string = '';
+  class: string = '';
+  isSortable: boolean = false;
+  isEditable: boolean = true;
+  isFilterable: boolean = false;
+  sortDirection: string = '';
+  defaultSortDirection: string = '';
+  editor: { type: string, config: any, component: any } = { type: '', config: {}, component: null };
+  filter: { type: string, config: any } = { type: '', config: {} };
+  renderComponent: any = null;
+  compareFunction: Function;
+  valuePrepareFunction: Function;
+  filterFunction: Function;
 
   constructor(public id: string, protected settings: any, protected dataSet: DataSet) {
     this.process();
   }
 
-  public getCompareFunction(): Function {
+  getCompareFunction(): Function {
     return this.compareFunction;
   }
 
-  public getValuePrepareFunction(): Function {
+  getValuePrepareFunction(): Function {
     return this.valuePrepareFunction;
   }
 
-  public getFilterFunction(): Function {
+  getFilterFunction(): Function {
     return this.filterFunction;
   }
 
-  public getCellRenderFunction(): Function {
-    return this.cellRenderFunction;
+  getConfig(): any {
+    return this.editor && this.editor.config;
   }
 
-  protected process(): void {
+  getFilterType(): any {
+    return this.filter && this.filter.type;
+  }
+
+  getFilterConfig(): any {
+    return this.filter && this.filter.config;
+  }
+
+  protected process() {
     this.title = this.settings['title'];
     this.class = this.settings['class'];
     this.type = this.prepareType();
+    this.editor = this.settings['editor'];
+    this.filter = this.settings['filter'];
+    this.renderComponent = this.settings['renderComponent'];
 
     this.isFilterable = typeof this.settings['filter'] === 'undefined' ? true : !!this.settings['filter'];
-    this.defaultSortDirection = ['asc', 'desc'].indexOf(this.settings['sortDirection']) !== -1 ? this.settings['sortDirection'] : '';
+    this.defaultSortDirection = ['asc', 'desc']
+      .indexOf(this.settings['sortDirection']) !== -1 ? this.settings['sortDirection'] : '';
     this.isSortable = typeof this.settings['sort'] === 'undefined' ? true : !!this.settings['sort'];
     this.isEditable = typeof this.settings['editable'] === 'undefined' ? true : !!this.settings['editable'];
     this.sortDirection = this.prepareSortDirection();
@@ -49,19 +63,18 @@ export class Column {
     this.compareFunction = this.settings['compareFunction'];
     this.valuePrepareFunction = this.settings['valuePrepareFunction'];
     this.filterFunction = this.settings['filterFunction'];
-    this.cellRenderFunction = this.settings['cellRenderFunction'];
   }
 
-  protected prepareType(): string {
+  prepareType(): string {
     return this.settings['type'] || this.determineType();
   }
 
-  protected prepareSortDirection(): string {
+  prepareSortDirection(): string {
     return this.settings['sort'] === 'desc' ? 'desc' : 'asc';
   }
 
-  protected determineType(): string {
+  determineType(): string {
     // TODO: determine type by data
-    return 'string';
+    return 'text';
   }
 }