+++ src/ng2-smart-table/lib/grid.ts
@@ -1,48 +1,54 @@
+import { Subject } from 'rxjs/Subject';
+import { Observable } from 'rxjs/Observable';
+import { EventEmitter } from '@angular/core';
+
+import { Deferred, getDeepFromObject } from './helpers';
 import { Column } from './data-set/column';
 import { Row } from './data-set/row';
 import { DataSet } from './data-set/data-set';
 import { DataSource } from './data-source/data-source';
-import { Subject, Observable } from 'rxjs/Rx';
-import { EventEmitter } from '@angular/core';
-import { Deferred, getDeepFromObject } from './helpers';
 
 export class Grid {
 
   createFormShown: boolean = false;
 
-  protected source: DataSource;
-  protected settings: any;
-  protected dataSet: DataSet;
+  source: DataSource;
+  settings: any;
+  dataSet: DataSet;
 
-  protected onSelectRowSource = new Subject<any>();
+  onSelectRowSource = new Subject<any>();
 
-  constructor(source: DataSource, settings) {
+  constructor(source: DataSource, settings: any) {
     this.setSettings(settings);
     this.setSource(source);
   }
-  
+
   showActionColumn(position: string): boolean {
     return this.isCurrentActionsPosition(position) && this.isActionsVisible();
   }
 
   isCurrentActionsPosition(position: string): boolean {
-    return position == this.getSetting('actions.position')
+    return position == this.getSetting('actions.position');
   }
 
   isActionsVisible(): boolean {
     return this.getSetting('actions.add') || this.getSetting('actions.edit') || this.getSetting('actions.delete');
   }
 
+  isMultiSelectVisible(): boolean {
+    return this.getSetting('selectMode') === 'multi';
+  }
+
   getNewRow(): Row {
     return this.dataSet.newRow;
   }
 
-  setSettings(settings: Object): void {
+  setSettings(settings: Object) {
     this.settings = settings;
     this.dataSet = new DataSet([], this.getSetting('columns'));
 
     if (this.source) {
-      this.source.refresh();  
+      this.source.refresh();
     }
   }
 
@@ -50,17 +56,17 @@ export class Grid {
     return this.dataSet;
   }
 
-  setSource(source: DataSource): void {
+  setSource(source: DataSource) {
     this.source = this.prepareSource(source);
 
     this.source.onChanged().subscribe((changes) => this.processDataChange(changes));
 
     this.source.onUpdated().subscribe((data) => {
-      let changedRow = this.dataSet.findRowByData(data);
+      const changedRow = this.dataSet.findRowByData(data);
       changedRow.setData(data);
     });
   }
-  
+
   getSetting(name: string, defaultValue?: any): any {
     return getDeepFromObject(this.settings, name, defaultValue);
   }
@@ -73,27 +79,31 @@ export class Grid {
     return this.dataSet.getRows();
   }
 
-  selectRow(row: Row): void {
+  selectRow(row: Row) {
     this.dataSet.selectRow(row);
+  }
+
+  multipleSelectRow(row: Row) {
+    this.dataSet.multipleSelectRow(row);
   }
 
   onSelectRow(): Observable<any> {
     return this.onSelectRowSource.asObservable();
   }
 
-  edit(row: Row): void {
+  edit(row: Row) {
     row.isInEditing = true;
   }
-  
-  create(row: Row, confirmEmitter: EventEmitter<any>): void {
 
-    let deferred = new Deferred();
+  create(row: Row, confirmEmitter: EventEmitter<any>) {
+
+    const deferred = new Deferred();
     deferred.promise.then((newData) => {
       newData = newData ? newData : row.getNewData();
       this.source.prepend(newData).then(() => {
         this.createFormShown = false;
         this.dataSet.createNewRow();
-      })
+      });
     }).catch((err) => {
       // doing nothing
     });
@@ -102,21 +112,21 @@ export class Grid {
       confirmEmitter.emit({
         newData: row.getNewData(),
         source: this.source,
-        confirm: deferred
+        confirm: deferred,
       });
     } else {
       deferred.resolve();
     }
   }
-  
-  save(row: Row, confirmEmitter: EventEmitter<any>): void {
 
-    let deferred = new Deferred();
+  save(row: Row, confirmEmitter: EventEmitter<any>) {
+
+    const deferred = new Deferred();
     deferred.promise.then((newData) => {
       newData = newData ? newData : row.getNewData();
       this.source.update(row.getData(), newData).then(() => {
         row.isInEditing = false;
-      })
+      });
     }).catch((err) => {
       // doing nothing
     });
@@ -126,55 +136,58 @@ export class Grid {
         data: row.getData(),
         newData: row.getNewData(),
         source: this.source,
-        confirm: deferred
+        confirm: deferred,
       });
     } else {
       deferred.resolve();
     }
   }
 
-  delete(row: Row, confirmEmitter: EventEmitter<any>): void {
-    
-    let deferred = new Deferred();
+  delete(row: Row, confirmEmitter: EventEmitter<any>) {
+
+    const deferred = new Deferred();
     deferred.promise.then(() => {
       this.source.remove(row.getData());
     }).catch((err) => {
       // doing nothing
     });
-    
+
     if (this.getSetting('delete.confirmDelete')) {
       confirmEmitter.emit({
         data: row.getData(),
         source: this.source,
-        confirm: deferred
+        confirm: deferred,
       });
     } else {
       deferred.resolve();
     }
   }
-  
-  protected processDataChange(changes): void {
+
+  processDataChange(changes: any) {
     if (this.shouldProcessChange(changes)) {
       this.dataSet.setData(changes['elements']);
-      let row = this.determineRowToSelect(changes);
-      if (row) {
-        this.onSelectRowSource.next(row);
+      if (this.getSetting('selectMode') !== 'multi') {
+        const row = this.determineRowToSelect(changes);
+
+        if (row) {
+          this.onSelectRowSource.next(row);
+        }
       }
     }
   }
-  
-  protected shouldProcessChange(changes): boolean {
+
+  shouldProcessChange(changes: any): boolean {
     if (['filter', 'sort', 'page', 'remove', 'refresh', 'load', 'paging'].indexOf(changes['action']) !== -1) {
       return true;
     } else if (['prepend', 'append'].indexOf(changes['action']) !== -1 && !this.getSetting('pager.display')) {
       return true;
     }
-    
+
     return false;
   }
 
   // TODO: move to selectable? Separate directive
-  protected determineRowToSelect(changes): Row {
+  determineRowToSelect(changes: any): Row {
 
     if (['load', 'page', 'filter', 'sort', 'refresh'].indexOf(changes['action']) !== -1) {
       return this.dataSet.select();
@@ -204,8 +217,8 @@ export class Grid {
     return null;
   }
 
-  protected prepareSource(source): DataSource {
-    let initialSource = this.getInitialSort();
+  prepareSource(source: any): DataSource {
+    const initialSource: any = this.getInitialSort();
     if (initialSource && initialSource['field'] && initialSource['direction']) {
       source.setSort([initialSource], false);
     }
@@ -217,8 +230,8 @@ export class Grid {
     return source;
   }
 
-  protected getInitialSort() {
-    let sortConf = {};
+  getInitialSort() {
+    const sortConf: any = {};
     this.getColumns().forEach((column: Column) => {
       if (column.isSortable && column.defaultSortDirection) {
         sortConf['field'] = column.id;
@@ -227,5 +240,16 @@ export class Grid {
       }
     });
     return sortConf;
+  }
+
+  getSelectedRows(): Array<any> {
+    return this.dataSet.getRows()
+      .filter(r => r.isSelected)
+      .map(r => r.getData());
+  }
+
+  selectAllRows(status: any) {
+    this.dataSet.getRows()
+      .forEach(r => r.isSelected = status);
   }
 }