+++ src/ng2-smart-table/lib/index.ts
@@ -1,4 +1,0 @@
-export * from './grid';
-export * from './data-source/data-source';
-export * from './data-source/local/local.data-source';
-export * from './data-source/server/server.data-source'